server/mcphub/memory/graphiti/requirements.txt
# requirements.txt
graphiti-core[google-genai]
python-dotenv
ipywidgets
langgraph
langchain-google-genai
python-dotenv
sentence-transformers